Modificado Repositorio de item. Modificado servicio de category e item. Modificado populate. Modificado ficheros spring, converter, message, security y tiles. Modificado header de master-page.
Añadidas vistas de category e item.
Añadido controlador de administrator con category e item
Añadido controlador global de item
Añadidos convertidores de category e item.

+++ Acme Veterinary/src/main/java/services/CategoryService.java
@@ -9,6 +9,7 @@ import org.springframework.util.Assert;
 
 import repositories.CategoryRepository;
 import domain.Category;
+import domain.Item;
 
 @Service
 @Transactional
@@ -21,6 +22,11 @@ public class CategoryService {
 
 	// Supporting services ----------------------------------------------------
 
+	@Autowired
+	private ActorService actorService;
+	@Autowired
+	private ItemService itemService;
+
 	// Constructors -----------------------------------------------------------
 
 	public CategoryService() {
@@ -30,31 +36,50 @@ public class CategoryService {
 	// Simple CRUD methods ----------------------------------------------------
 
 	public Category create() {
+		Assert.isTrue(actorService.isAdministrator());
 		Category result;
 		result = new Category();
 		return result;
 	}
 
-	public Category findOne(int categoryId) {
-		Category result;
-		result = categoryRepository.findOne(categoryId);
+	public Category save(Category category) {
+		Assert.isTrue(actorService.isAdministrator());
+		Assert.notNull(category);
+		return categoryRepository.save(category);
+	}
+
+	public boolean delete(Category category) {
+		Assert.isTrue(actorService.isAdministrator());
+		Assert.notNull(category);
+		boolean result;
+		Collection<Item> items;
+		items = itemService.findByCategory(category.getId());
+
+		if (items.size() == 0) {
+			categoryRepository.delete(category);
+			result = true;
+		} else {
+			result = false;
+		}
 		return result;
 	}
 
 	public Collection<Category> findAll() {
+		Assert.isTrue(actorService.isAdministrator());
 		Collection<Category> result;
 		result = categoryRepository.findAll();
 		return result;
 	}
 
-	public Category save(Category category) {
-		Assert.notNull(category);
-		return categoryRepository.save(category);
+	public Category findOne(int categoryId) {
+		Assert.isTrue(actorService.isAdministrator());
+		Assert.isTrue(categoryId != 0);
+		Category result;
+		result = categoryRepository.findOne(categoryId);
+		Assert.notNull(result);
+		return result;
 	}
-	
-	public void delete(Category category) {
-		categoryRepository.delete(category);
-	}
-	
-	// Other business methods -------------------------------------------------
+
+	// Other business methods
+
 }
